fix(resolver): wipe ScriptAssemblies in Force Resolve menu + extract shared util
The same stale-AppDomain problem the package update popup guards against
also bites the manual "Force Resolve NuGet DLLs" menu: when user-asmdef
compile errors prevent Unity from completing a domain reload, force-resolve
runs against the OLD plugin assemblies in memory and any new resolver code
on disk stays idle.

Pull the wipe helper out of UpdatePopupWindow (where I had it as a private
method) into a shared NuGetExtractor.DependencyResolver.ScriptAssembliesCache
utility, and call it at the top of NuGetResolverMenu.ForceResolve. The
update popup now delegates to the same helper.
